portfolio_projects/data_engineering/kafka_projects/stock_market_project/requirements.txt
confluent-kafka
apache-airflow-providers-apache-kafka==1.7.0
apache-airflow-providers-amazon>=9.2.0
apache-airflow-providers-ssh
boto3
s3fs>=2023.1.0
pandas>=1.3.5
requests>=2.31.0
python-dotenv>=1.0.0
six